+++ src/Knp/Component/Pager/Event/Subscriber/Filtration/FiltrationSubscriber.php
@@ -22,8 +22,8 @@ class FiltrationSubscriber implements EventSubscriberInterface
         /** @var \Symfony\Component\EventDispatcher\EventDispatcherInterface $dispatcher */
         $dispatcher = $event->getEventDispatcher();
         // hook all standard filtration subscribers
-        $dispatcher->addSubscriber(new Doctrine\ORM\QuerySubscriber($event->getRequest()));
-        $dispatcher->addSubscriber(new PropelQuerySubscriber($event->getRequest()));
+        $dispatcher->addSubscriber(new Doctrine\ORM\QuerySubscriber($event->getArgumentAccess()));
+        $dispatcher->addSubscriber(new PropelQuerySubscriber($event->getArgumentAccess()));
 
         $this->isLoaded = true;
     }