+++ src/Knp/Component/Pager/Event/Subscriber/Sortable/Doctrine/ORM/QuerySubscriber.php
@@ -3,28 +3,25 @@
 namespace Knp\Component\Pager\Event\Subscriber\Sortable\Doctrine\ORM;
 
 use Doctrine\ORM\Query;
+use Knp\Component\Pager\ArgumentAccess\ArgumentAccessInterface;
 use Knp\Component\Pager\Event\ItemsEvent;
 use Knp\Component\Pager\Event\Subscriber\Paginate\Doctrine\ORM\Query\Helper as QueryHelper;
 use Knp\Component\Pager\Event\Subscriber\Sortable\Doctrine\ORM\Query\OrderByWalker;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpFoundation\Request;
 
 class QuerySubscriber implements EventSubscriberInterface
 {
-    /**
-     * @var Request
-     */
-    private $request;
+    private ArgumentAccessInterface $argumentAccess;
 
-    public function __construct(Request $request)
+    public function __construct(ArgumentAccessInterface $argumentAccess)
     {
-        $this->request = $request;
+        $this->argumentAccess = $argumentAccess;
     }
 
     public function items(ItemsEvent $event): void
     {
-        // Check if the result has already been sorted by an other sort subscriber
+        // Check if the result has already been sorted by another sort subscriber
         $customPaginationParameters = $event->getCustomPaginationParameters();
         if (!empty($customPaginationParameters['sorted']) ) {
             return;
@@ -34,16 +31,16 @@ class QuerySubscriber implements EventSubscriberInterface
             $event->setCustomPaginationParameter('sorted', true);
             $sortField = $event->options[PaginatorInterface::SORT_FIELD_PARAMETER_NAME];
             $sortDir = $event->options[PaginatorInterface::SORT_DIRECTION_PARAMETER_NAME];
-            if (null !== $sortField && $this->request->query->has($sortField)) {
-                $dir = null !== $sortDir && $this->request->query->has($sortDir) && strtolower($this->request->query->get($sortDir)) === 'asc' ? 'asc' : 'desc';
+            if (null !== $sortField && $this->argumentAccess->has($sortField)) {
+                $dir = null !== $sortDir && $this->argumentAccess->has($sortDir) && strtolower($this->argumentAccess->get($sortDir)) === 'asc' ? 'asc' : 'desc';
 
                 if (isset($event->options[PaginatorInterface::SORT_FIELD_ALLOW_LIST])) {
-                    if (!in_array($this->request->query->get($sortField), $event->options[PaginatorInterface::SORT_FIELD_ALLOW_LIST])) {
-                        throw new \UnexpectedValueException("Cannot sort by: [{$this->request->query->get($sortField)}] this field is not in allow list.");
+                    if (!in_array($this->argumentAccess->get($sortField), $event->options[PaginatorInterface::SORT_FIELD_ALLOW_LIST])) {
+                        throw new \UnexpectedValueException("Cannot sort by: [{$this->argumentAccess->get($sortField)}] this field is not in allow list.");
                     }
                 }
 
-                $sortFieldParameterNames = $this->request->query->get($sortField);
+                $sortFieldParameterNames = $this->argumentAccess->get($sortField);
                 $fields = [];
                 $aliases = [];
                 if (!is_string($sortFieldParameterNames)) {