+++ src/Knp/Component/Pager/Event/Subscriber/Sortable/ElasticaQuerySubscriber.php
@@ -4,23 +4,23 @@ namespace Knp\Component\Pager\Event\Subscriber\Sortable;
 
 use Elastica\Query;
 use Elastica\SearchableInterface;
+use Knp\Component\Pager\ArgumentAccess\ArgumentAccessInterface;
 use Knp\Component\Pager\Event\ItemsEvent;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpFoundation\Request;
 
 class ElasticaQuerySubscriber implements EventSubscriberInterface
 {
-    private Request $request;
+    private ArgumentAccessInterface $argumentAccess;
 
-    public function __construct(Request $request)
+    public function __construct(ArgumentAccessInterface $argumentAccess)
     {
-        $this->request = $request;
+        $this->argumentAccess = $argumentAccess;
     }
 
     public function items(ItemsEvent $event): void
     {
-        // Check if the result has already been sorted by an other sort subscriber
+        // Check if the result has already been sorted by another sort subscriber
         $customPaginationParameters = $event->getCustomPaginationParameters();
         if (!empty($customPaginationParameters['sorted']) ) {
             return;
@@ -31,9 +31,9 @@ class ElasticaQuerySubscriber implements EventSubscriberInterface
             $event->setCustomPaginationParameter('sorted', true);
             $sortField = $event->options[PaginatorInterface::SORT_FIELD_PARAMETER_NAME];
             $sortDir = $event->options[PaginatorInterface::SORT_DIRECTION_PARAMETER_NAME];
-            if (null !== $sortField && $this->request->query->has($sortField)) {
-                $field = $this->request->query->get($sortField);
-                $dir   = null !== $sortDir && $this->request->query->has($sortDir) && strtolower($this->request->query->get($sortDir)) === 'asc' ? 'asc' : 'desc';
+            if (null !== $sortField && $this->argumentAccess->has($sortField)) {
+                $field = $this->argumentAccess->get($sortField);
+                $dir   = null !== $sortDir && $this->argumentAccess->has($sortDir) && strtolower($this->argumentAccess->get($sortDir)) === 'asc' ? 'asc' : 'desc';
 
                 if (isset($event->options[PaginatorInterface::SORT_FIELD_ALLOW_LIST]) && !in_array($field, $event->options[PaginatorInterface::SORT_FIELD_ALLOW_LIST])) {
                     throw new \UnexpectedValueException(sprintf('Cannot sort by: [%s] this field is not in allow list.', $field));