+++ tests/Test/Pager/Subscriber/Paginate/ArrayTest.php
@@ -2,6 +2,7 @@
 
 namespace Test\Pager\Subscriber\Paginate;
 
+use Knp\Component\Pager\ArgumentAccess\ArgumentAccessInterface;
 use Knp\Component\Pager\Event\Subscriber\Paginate\ArraySubscriber;
 use Knp\Component\Pager\Pagination\PaginationInterface;
 use Knp\Component\Pager\Paginator;
@@ -19,7 +20,8 @@ final class ArrayTest extends BaseTestCase
         $dispatcher = new EventDispatcher;
         $dispatcher->addSubscriber(new ArraySubscriber);
         $dispatcher->addSubscriber(new MockPaginationSubscriber); // pagination view
-        $p = new Paginator($dispatcher);
+        $accessor = $this->createMock(ArgumentAccessInterface::class);
+        $p = new Paginator($dispatcher, $accessor);
 
         $items = ['first', 'second'];
         $view = $p->paginate($items, 1, 10);
@@ -39,7 +41,8 @@ final class ArrayTest extends BaseTestCase
         $dispatcher = new EventDispatcher;
         $dispatcher->addSubscriber(new ArraySubscriber);
         $dispatcher->addSubscriber(new MockPaginationSubscriber); // pagination view
-        $p = new Paginator($dispatcher);
+        $accessor = $this->createMock(ArgumentAccessInterface::class);
+        $p = new Paginator($dispatcher, $accessor);
 
         $items = \range('a', 'u');
         $view = $p->paginate($items, 2, 10);