+++ tests/Test/Pager/Subscriber/Paginate/Doctrine/CollectionTest.php
@@ -3,6 +3,7 @@
 namespace Test\Pager\Subscriber\Paginate\Doctrine;
 
 use Doctrine\Common\Collections\ArrayCollection;
+use Knp\Component\Pager\ArgumentAccess\ArgumentAccessInterface;
 use Knp\Component\Pager\Event\Subscriber\Paginate\Doctrine\CollectionSubscriber;
 use Knp\Component\Pager\Pagination\PaginationInterface;
 use Knp\Component\Pager\Paginator;
@@ -20,7 +21,8 @@ final class CollectionTest extends BaseTestCase
         $dispatcher = new EventDispatcher;
         $dispatcher->addSubscriber(new CollectionSubscriber);
         $dispatcher->addSubscriber(new MockPaginationSubscriber); // pagination view
-        $p = new Paginator($dispatcher);
+        $accessor = $this->createMock(ArgumentAccessInterface::class);
+        $p = new Paginator($dispatcher, $accessor);
 
         $items = new ArrayCollection(['first', 'second']);
         $view = $p->paginate($items, 1, 10);
@@ -39,7 +41,8 @@ final class CollectionTest extends BaseTestCase
         $dispatcher = new EventDispatcher;
         $dispatcher->addSubscriber(new CollectionSubscriber);
         $dispatcher->addSubscriber(new MockPaginationSubscriber); // pagination view
-        $p = new Paginator($dispatcher);
+        $accessor = $this->createMock(ArgumentAccessInterface::class);
+        $p = new Paginator($dispatcher, $accessor);
 
         $items = new ArrayCollection(['first', 'second']);
         $view = $p->paginate($items, 1, 10);
@@ -63,7 +66,8 @@ final class CollectionTest extends BaseTestCase
         $dispatcher = new EventDispatcher;
         $dispatcher->addSubscriber(new CollectionSubscriber);
         $dispatcher->addSubscriber(new MockPaginationSubscriber); // pagination view
-        $p = new Paginator($dispatcher);
+        $accessor = $this->createMock(ArgumentAccessInterface::class);
+        $p = new Paginator($dispatcher, $accessor);
 
         $items = new ArrayCollection(\range('a', 'u'));
         $view = $p->paginate($items, 2, 10);